requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
alembic
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.0.1
python-multipart
cloudinary
python-dotenv
httpx
qrcode[pil]
Pillow
pydantic-settings
